Extract immersive mode plugin
- Extracted MediaViewer.js, ImmersiveSheetViewer.js, and gestures.js into frontend/src/plugins/immersive
- Updated all relative imports to point to new locations
- Created immersive plugin index.js mounting function
- Removed immersive_overlay_mode global setting, replaced with plugin.immersive.mode
- Verified gestures tests pass and code bundles correctly via esbuild
