+++ api/internal/api/settings.go
@@ -38,7 +38,7 @@ var publicSettingKeys = map[string]bool{
 	"tags_visibility":         true,
 	"enable_backup":           true,
 	"immersive_nav_direction": true,
-	"immersive_overlay_mode":  true,
+	"plugin.immersive.mode": true,
 	"exif_visibility":         true,
 	"nav_menu_mode":           true,
 }
