+++ frontend/src/plugins/immersive/MediaViewer.js
@@ -20,11 +20,11 @@
  *   navNext    {object|null}  { slug, title }
  */
 
-import { Component } from '../Component.js';
+import { Component } from '../../components/Component.js';
 import { escapeHtml, safeUrl, sharePost, navigate } from '../../utils/helpers.js';
 import { SHARE_SVG } from '../../utils/icons.js';
 import { store } from '../../store.js';
-import { GestureController, TrackpadDetector, rubberBand } from '../../utils/gestures.js';
+import { GestureController, TrackpadDetector, rubberBand } from './gestures.js';
 import { hideFlyout } from '../../utils/tags.js';
 import { ViewContext } from '../../utils/viewContext.js';
 import { getPostBySlug, getPostNavigation } from '../../api/posts.js';
