+++ frontend/src/plugins/immersive/gestures.js
